requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
html2text==2024.2.26
pyyaml==6.0.2
lxml==5.3.0
markdownify==0.13.1
rich==13.8.1
click==8.1.7
selenium==4.25.0
webdriver-manager==4.0.2
python-dotenv==1.0.1